BACKPORT: drm/tegra: dpaux: Support monitor hotplugging
The dpaux driver has a quirk built-in that will delay initialization of
the display driver for a short while, trying to detect an eDP panel. The
reason for this quirk is that the panel may not report as connected
until after the display driver has initialized, at which point the fbdev
emulation will have fallen back to 1024x768 as default resolution, which
will likely not be the eDP panel's native resolution.

With upcoming DisplayPort support, the code needs to be able to cope
with hotpluggable monitors as well. Waiting for a panel to show up is no
longer going to work because the monitor may not be attached on boot. If
the output runs in DisplayPort mode, skip waiting for the panel to show
up.

Signed-off-by: Thierry Reding <treding@nvidia.com>
[from https://github.com/thierryreding/linux.git]
(cherry picked from commit 503d9f806250690d48aa0ffef83e09d293a87236)
[seanpaul fixed wait_for conflicts]
Signed-off-by: Sean Paul <seanpaul@chromium.org>

Change-Id: Id9fe20b00fd5a244a172dc10e42f7c98f073d4c6
Original-Change-Id: Iac47fa7dd7489b3db7d52c0e1935866be9ebd1b8
Reviewed-on: https://chromium-review.googlesource.com/298349
Reviewed-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>