FROMLIST: drm: bridge/dw_hdmi: remove ratio support from ACR code
We never set the ratio for CTS/N calculation for the audio clock
regenerator (ACR) to anything but 100, so this adds pointless
complexity.  Should we support pixel repetition, we should update the
CTS/N calculation code to use those parameters or the actual TMDS clock
rate instead of a ratio.

Conflicts:
    drivers/gpu/drm/bridge/dw_hdmi.c
We appeared to have an earlier version of patches.

BUG=chrome-os-partner:44695
TEST=Audio works w/ lots of resolutions

Change-Id: I7a149b427552c56e7a10522ba41f9fdcf6943655
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/6975241/)
Reviewed-on: https://chromium-review.googlesource.com/297617
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>