CHROMIUM: ARM: dts: Try harder to keep GPU at 400MHz and CPU at 816MHz
Apparently bringing the GPU down to 300MHz is not so great.  Let's try
a little harder to keep it at 400 by throttling the CPU a little more.

We'll also try to keep the CPU at 816 at least, since that appears to
be a sweet spot that we can maintain nicely.

Note that we'll bump the GPU throttling temps 1C above the CPU just
to try extra hard not to throttle the GPU.

BUG=chrome-os-partner:42854
TEST=Run watchspeed while doing webgl aquarium

Change-Id: I6b11ba27f36721303598af74d323f1315cef41b4
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/293946
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
(cherry picked from commit 07e8442ea03ae35d4783df5d340fd3c986997b57)
Reviewed-on: https://chromium-review.googlesource.com/294756