CHROMIUM: clk: rockchip: Add high Fvco 308 MHz rate for 154 MHz HDMI
We'd like to reduce jitter for 154 MHz HDMI clock rates.  Although it
is suggested to keep Fvco under 2.2 GHz, the actual listed max is 5
GHz and we actually use something close to that rate for 594 MHz.

We'll copy what was done for 594MHz here and also adjust the PLL
bandwidth.  Together the high Fvco and low bwadj give us a very low
jitter clock according to measurements.

BUG=chrome-os-partner:43781
TEST=Measure jitter; try on real monitor/TV

Change-Id: Ib0164513bd3d75196b590753f6a9a6eea990c49a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/294333