CHROMIUM: clk: rockchip: Add high Fvco 292.5 MHz rate for 146.25 MHz HDMI
We'd like to reduce jitter for 146.25 MHz HDMI clock rates.  Although it
is suggested to keep Fvco under 2.2 GHz, the actual listed max is 5
GHz and we actually use something close to that rate for 594 MHz.

We'll copy what was done for 594MHz here and also adjust the PLL
bandwidth.  Together the high Fvco and low bwadj give us a very low
jitter clock according to measurements.

BUG=chrome-os-partner:43781
TEST=Measure jitter; try on real monitor/TV

Change-Id: Ib30517b70361d337c4f83bab51fe0551f339cfa7
83;40201;0cSigned-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/294334
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>