CHROMIUM: drm: rockchip/dw_hdmi-rockchip: Better 119 and 136.75
We've added low jitter PLL rates for 273.6 MHz and 238 MHz.  Let's use
them to better make 136.75 MHz and 119 MHz.

Note that we'll leave the 118.67 rate in place because it appears to
be closer than 119 with at least one mode setting: "1920x1080@ 48Hz,
clk=118680" and is jitter has already been validated.

BUG=chrome-os-partner:43781
TEST=Measure jitter; try on real monitor/TV

Change-Id: I8e2bbc0bb5059b28211ea809e076a38f39456a62
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295304