CHROMIUM: clk: rockchip: Add 160 MHz for 32 MHz HDMI
Add a 160 MHz clock that will be useful for making a 32 MHz pixel
clock for HDMI (divide it by 5).  The settings used has and nr of 1
and an Fvco of 1.92 GHz, both things are good for jitter (not that
jitter matters terribly at 32 MHz).

If we don't add this we could still make 32 MHz but the clock
framework will pick 384 MHz and our current 384 MHz clock rate seems
like it won't be quite as good for jitter (it has an nr of 2 and Fvco
of 1.536 GHz).

BUG=chrome-os-partner:44512
TEST=Screen works

Change-Id: I7e3fd22964e0b85231e6190ac933a5e67e97268e
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295541
Reviewed-by: Yakir Yang <ykk@rock-chips.com>