CHROMIUM: soc/tegra: pmc: Complete powergate sequence
This patch modifies the powergate sequence to make it
as completed for all power partitions.

For powergate power down, the sequence should be:
 - Enable all related clocks
 - Set related MC clients FLUSH_ENABLE bit
 - Assert all reset devices
 - Disable all related clocks
 - Set PMC powergate toggle for specified partition

For powergate power up, the sequence should be:
 - Set PMC powergate toggle for specified partition
 - Enable all related clocks
 - Assert all reset devices
 - Remove power gating clamps
 - Deassert all reset devices
 - Clear related MC clients FLUSH_ENABLE bit
 - Disable all related clocks

So, two things will be done in this patch:
1. Add mc_flush support
2. Fix powergate sequence

BTW, the mc driver data is stored for mc_flush using,
so it also moves the tegra_powergate_init() into
tegra_pmc_probe() to ensure tegra_powergate_init() is
invoked after mc driver is ready.

Powergate sequence is based on the work by:
Mayuresh Kulkarni <mkulkarni@nvidia.com>

BUG=chrome-os-partner:44838
TEST=Build and boot smaug

Change-Id: Id388704c5dc8436c46a360a5fa9ff51be845e693
Signed-off-by: Penny Chiu <pchiu@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/298371
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>