Revert "CHROMIUM: drm: rockchip/hdmi: switch ddc to ip ddc integrated module"
This reverts commit 58dadb6461f4f2ca8f41b4950f2836cd03e093c0.

That commit had a few big problems:
1. It changed the DTS and code at the same time, which we try to avoid
   as much as possible to help w/ future merges.
2. The C code whacks the pinmux directly.  That's not a good idea both
   for general cleanliness and because it's touching bits managed by
   another driver.

In fact point #2 seems to be causing real problems.  On Mickey I can no
longer read the EDID over HDMI.  That's because HDMI happens to probe
before i2c5 on Mickey and thus when we load up i2c5 (still says 'status
= "okay" in DTS!) we'll clobber the pinmux.

The cleanest thing is to just revert the bad change and put in a proper
change, which should be DTS only.

BUG=chrome-os-partner:44809
TEST=Mickey can boot up again and read EDID properly

Change-Id: I7da236d655fa2e96f5f7230a8d237037e1e2b36e
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/297009
Reviewed-by: Alexandru Stan <amstan@chromium.org>