CHROMIUM: drm: bridge/dw_hdmi: Add special cases for N for >= 297 MHz
There appear to be some (unknown) reasons for keeping special cases of N
for rates >= 297 MHz in some cases.  Until we can figure out what those
reasons are, let's add back special cases for 297 MHz (and 594 MHz from
the Rockchip datasheet while we're at it).

Note that we don't add special cases for non-exact rates.  The default
forumla will have to be good enough for that.

BUG=chrome-os-partner:44695
TEST=Audio works w/ lots of resolutions

Change-Id: I99d322af08855ffda4b6e322edad7660a47b0368
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/297790
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>