CHROMIUM: drm: bridge/dw_hdmi: Add "unwedge" for dw_hdmi i2c bus
See the PhD thesis in the comments in this patch for details, but to
summarize this adds a hacky "unwedge" feature to the dw_hdmi i2c bus to
workaround what appears to be a hardware errata.  This relies on a
pinctrl entry to help change around muxing to perform the unwedge.

BUG=chrome-os-partner:45755
TEST=With series unwedge works now; see printout in logs.

Change-Id: I977adbbce0bd00199f7eab9a6dfe17618ca31baa
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/299493
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>