DROP: Hack: CHROMIUM: usb: gadget: tegra_xudc: Disable U2
There is a U1->U2 inactivity timer in HW which is active in U1 state,
and the state machine automatically transitions to U2 when that timer
expires. Ideally this timer should be reset and stopped when U1 exit is
in progress; this is not happening, which causes the state machine to
abruptly change to U2 when U1 exit LFPS handshake is in progress.

This known issue will happen when transferring large files (e.g. 3.8GB)
from some XHCI port to device where incorrect state machine transition
is observed: U0->Inactive->RxDetect->U0.

So we're not able to enable both U1 and U2.
In https://chromium-review.googlesource.com/299563, the decision is to
disable U1 only since major power saving comes from U2.

For factory environment, we want to do the opposite - get faster
response and use more power.

BUG=none
TEST=Smaug XUSB device mode works normally

Change-Id: I6b3a23a1c2d8f01e1769be2073fb817e653c2f24
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/307691
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>