FROMLIST: ath10k: fix ldpc vdev param for fixed rate
ldpc is disabled when fixed rate is used. As firmware is considering ldpc vdev
param for fixed rate selection, ldpc is disabled when it is not configured.
This vdev param is ignored if client does not support ldpc at all.

Signed-off-by: Rajkumar Manoharan <rmanohar@qti.qualcomm.com>
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>

BUG=chrome-os-partner:44887
TEST=build/boot on storm suceeded. Setup Storm board as AP using hostapd and
connected to the board using another device. Device was able to connect to
the internet and load multiple websites.

Change-Id: Ic05def5d25ba913f9ded7ef665b56e831c2f5ea6
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/299598
Commit-Ready: Yixiang Li <yixiang@google.com>
Tested-by: Yixiang Li <yixiang@google.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: Yixiang Li <yixiang@google.com>