CHROMIUM: go2001: Pass allow_zero_bytesused = 1 to vb2_queue_init().
go2001 allows queuing bytesused = 0 OUTPUT buffers as a way to request
decoder flush. This is a legacy/nonstandard way of doing this, so we
need to set allow_zero_bytesused = 1 when calling vb2_queue_init() to
override the default behavior in vb2 to disallow bytesused = 0 buffers.

Signed-off-by: Pawel Osciak <posciak@chromium.org>

BUG=None
TEST=video playback with flush

Change-Id: I8e88b7526c5d37a01fce25657e3d78e73fc5c4cf
Reviewed-on: https://chromium-review.googlesource.com/300724
Commit-Ready: Pawel Osciak <posciak@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Owen Lin <owenlin@chromium.org>