CHROMIUM: drm: bridge/dw_hdmi: Fix up HPD interrupt handling
In hdmi_unmute_interrupts() we were always re-initting the HPD polarity
to detect insertions, not removals.  This wasn't a huge deal because
hdmi_unmute_interrupts() was only called at init time and resume time.
At init time we only need to detect insertions and at resume time
skipping dw_hdmi_irq() seems to be OK even HDMI was unplugged while
asleep.  ...but it's non-ideal.

Let's fix hdmi_unmute_interrupts() to make sure it inits the HPD
polarity properly.  This will make sure we pass through dw_hdmi_irq() at
resume time if a cable is unplugged while asleep, which seems cleaner.
This is also prep work for calling hdmi_unmute_interrupts() in other
instances.

In order to accomplish the above, we need to introduce a new state
variable about whether we think the cable is plugged in or unplugged.
This needs to be in-sync with the polarity.  We introduce a mutex to
make sure they're in sync.  At the moment the mutex is a bit of an
overkill since it's unlikely that dw_hdmi_irq() and
hdmi_unmute_interrupts() are running at the same time (again, since
hdmi_unmute_interrupts() is only called at resume time right now), but
if/when we call hdmi_unmute_interrupts() from more places the mutex will
become more important.

BUG=chrome-os-partner:44922
TEST=Insert/remove HDMI, both while suspended and while awake.

Change-Id: I4d8214ed0c91249bc0054792fe8d432ec227b4b9
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/299948
Reviewed-by: Yakir Yang <ykk@rock-chips.com>