HACK: CHROMIUM: drm/edid: force LG eDP panels (for samus) to 6bpc
This hack should address (for now) some flickering issues seen on
certain panels found on samus. These issues only appear in developer
mode, where we run the VBIOS and (normally) end up configuring for 24
bpp (8bpc). In non-developer mode, we skip the VBIOS and due to a
firmware bug, end up using 18 bpp (6 bpc) anyway.

Now, this panel should support 24 bpp, so there is still more
investigation to be done, but for now, we might as well just use the
same settings in both developer mode and standard mode. When we can root
cause the issue, we can fix things up to switch both to 24 bpp.

BUG=chrome-os-partner:43992,chromium:498451
TEST=boot samus in developer mode, check for flicker (note: not all
     samus exhibit the flicker; just MP?)

Change-Id: I2307e24690162137c9465c1703966adad844201c
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/301173
Tested-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>