BACKPORT: ASoC: rt5645: Add jd_invert for Broadwell
Broadwell can not triger the IRQ falling and rising simultaneously,
so it can not detect jack-in and jack-out simultaneously.
We add a flag "jd_invert" to platform data. If this flag is set,
codec IRQ will be set to invert that forces IRQ as pulse when jack-in
and jack-out.

Signed-off-by: John Lin <john.lin@realtek.com>
Signed-off-by: Bard Liao <bardliao@realtek.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 917536aeb88d34e06c1353b0dd144f0987bb66bd)
Signed-off-by: Ben Zhang <benzh@chromium.org>

Conflicts:
	sound/soc/codecs/rt5645.c - context fix

BUG=chrome-os-partner:43037
TEST=4ch recording on Buddy with TDM

Change-Id: I2b834ce393465946286a729a4f289221885901ef
Reviewed-on: https://chromium-review.googlesource.com/300604
Commit-Ready: Ben Zhang <benzh@chromium.org>
Tested-by: Ben Zhang <benzh@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>