CHROMIUM: drm: bridge/dw_hdmi: Protect read/modify/write
The dw_hdmi driver had a "read-modify-write" function for changing
registers, but there was no locking around it.  Let's add a lock.

This was found only by code inspection.  No known bugs are fixed by
this.  May be important for HDCP patches which make changes to one of
these registers from a timer function.

BUG=chrome-os-partner:34741
TEST=Boot up; HDMI works OK.

Change-Id: Iaeb357562ff30f19f676b2a26e945fdfdadb4c04
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/302080
Tested-by: Yakir Yang <ykk@rock-chips.com>
Reviewed-by: Yakir Yang <ykk@rock-chips.com>