CHROMIUM: HACK: sound: usb: fix rk3288 usbspeaker play music staccato issue
When connect the usb speaker with hub to play music, the cpu frequency
may drop to 126M, the ddr frequency may scaling(which will hold the cpus 500us),
all of these may cause staccato music.
To avoid that, we set cpu scaling_min_freq to specific value(which define in dts)
and disable dmc when detect usb speaker plug in.

BUG=chrome-os-partner:43539
TEST=connect usb speaker with hub, play music, it work normally

Change-Id: I9153a48537d3a3a00f9c64252108b587725d34aa
Signed-off-by: Lin Huang <hl@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/302392
Reviewed-by: Douglas Anderson <dianders@chromium.org>