CHROMIUM: drm: bridge/dw_hdmi: Init i2c after resume
We need to make sure we are able to do i2c over the HDMI i2c port again
after suspend/resume, so we'll put an init in the resume function.  Note
that we refactor the initialize_hdmi_ih_mutes() a bit because at resume
time it's wise to init the i2c stuff before we init the rest of the
interrupts.  The initialize_hdmi_ih_mutes() was always duplicated at
bind time anyway...

BUG=chrome-os-partner:34741
TEST=HDMI vs. suspend/resume, broken monitor, HDCP, etc.

Change-Id: Idf0ded22d6731d97649c3129f70ee5f0b96894f4
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/302800
Commit-Ready: Yakir Yang <ykk@rock-chips.com>
Reviewed-by: Yakir Yang <ykk@rock-chips.com>