CHROMIUM: usb: dwc2: Avoid double-reset at boot time
In (5a87723 FROMLIST: usb: dwc2: reset dwc2 core before
dwc2_get_hwparams()) AKA CL:292610 we added an extra reset to the probe
path for the dwc2 USB controllers.  This allowed proper detection of
parameters even if the firmware had already used the USB part.

Unfortunately, this extra reset is quite slow and seems to be affecting
boot speed (which is a little strange since the probe should be
asynchronous?).  In any case, we can avoid the double-reset by skipping
the extra reset that would happen just after the one we added.  Logic
that explains why this is safe:

* As of the CL mentioned above, we now always call dwc2_core_reset() in
  dwc2_driver_probe() before dwc2_hcd_init().

* The only caller of dwc2_hcd_init() is dwc2_driver_probe(), so we're
  guaranteed that dwc2_core_reset() was called before dwc2_hdc_init().

* dwc2_hdc_init() is the only caller that passes an irq other than -1 to
  dwc2_core_init().  Thus if dwc2_core_init() is called with an irq
  other than -1 we're guaranteed that dwc2_core_reset was called before
  dwc2_core_init().

...this allows us to remove the dwc2_core_reset() in dwc2_core_init() if
irq is not < 0.

Note that since "irq" wasn't used in the function dwc2_core_init()
anyway and since select_phy was always set at exactly the same times we
could avoid the reset, we remove "irq" and rename "select_phy" to
"initial_setup" and adjust the callers accordingly.

BUG=chromium:530440
TEST=bootperf; USB works after bootup on both ports; no new USB warnings

Change-Id: I1a5bca4f568c209abcef1329b7fa40633696a139
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/303152
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>