CHROMIUM: move WIFIVERSION Kconfig to the right place
With this being in drivers/net/Kconfig, net/Kconfig will still
include net/wireless/Kconfig instead of net/wireless-3.8/Kconfig
and the same for mac80211.

This seems fairly unproblematic as the Kconfig files are almost
identical, for mac80211 we only lack the MAC80211_MESH_PS_DEBUG.

However, our driver is now relying on the WIRELESS_38 symbol,
which is never defined right now.

Fix this by moving the WIFIVERSION to the place where it's first
needed.

Change-Id: I6a724f0db6dce634f8f83197c8adfdfdad3b0163
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/303550
Reviewed-by: Anatol Pomazau <anatol@google.com>