CHROMIUM: drm: rockchip/hdmi: Fix suspend/resume problems
It looks as if the VOP gets resumed before HDMI does.  Then, it goes and
calls:
  rockchip_drm_sys_resume
  drm_helper_connector_dpms
  drm_helper_encoder_dpms
  dw_hdmi_bridge_enable
  dw_hdmi_setup

...this happens before the HDMI resume, which isn't such a great thing.

At some point we can try to think about how to make that all work
better, but for now let's take the quick fix of moving dw_hdmi to early
resume and then we're all good.

BUG=chrome-os-partner:46106
TEST=suspend_stress_test

Change-Id: Ibc61b5e4f9ece5083211624889cb9a49c6d021d2
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/303922
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>