UPSTREAM: drm/edid: Fix up clock for CEA/HDMI modes specified via detailed timings
EDID detailed timings have a resolution of 10kHz for the pixel clock, so
they can't represent certain CEA/HDMI modes accurately. If we see a mode
coming in via detailed timings which otherwise matches one of the
CEA/HDMI modes except the clock is just a bit off, let's assume that the
intention was for that mode to be one of the CEA/HDMI modes and go ahead
and fix up the clock to match the CEA/HDMI spec exactly (well, as close
as we can get with the 1 kHz resolution we use).

This should help code that's looking for an exact clock match (eg. i915
audio N/CTS setup).

Cc: Adam Jackson <ajax@redhat.com>
Cc: Clint Taylor <clinton.a.taylor@intel.com>
Cc: Libin Yang <libin.yang@intel.com>
Signed-off-by: Ville Syrjl <ville.syrjala@linux.intel.com>
Reviewed-by: Adam Jackson <ajax@redhat.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>

Link: git://anongit.freedesktop.org/drm-intel
(cherry picked from commit 7a1994136f3a0f32d0f80dff760d3c4d2ea3591b)
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>

BUG=chrome-os-partner:38919
TEST=On BDW, with entire patch series applied, verify
     correct programming of N value for TMDS_296M and
     TMDS_297M with 4K HDMI test instrument for multiple
     audio rates. Verify proper audio output on 4K HDMI
     displays.

Change-Id: I848a3a1c8c4989b681f964e19d29651d2f0ae328
Reviewed-on: https://chromium-review.googlesource.com/305511
Commit-Ready: Ullysses A Eoff <ullysses.a.eoff@intel.com>
Tested-by: Nathan D Ciobanu <nathan.d.ciobanu@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Clinton A Taylor <clinton.a.taylor@intel.com>