CHROMIUM: drm/tegra: Unpowergate DIS for dcs messages
Don't rely on the dc to have the DIS power domain unpowergated
since we might shut down while the panel is powered and the dc
is not active.

This patch adds an unpowergate/powergate pair around the dcs
transfer function in the dsi driver to ensure that the dsi is
ungated before writing registers. Since the powergate code is
refcounted, we can do this unconditionally.

BUG=chrome-os-partner:45928
TEST=Tested on smaug, do: fastboot erase userdata && fastboot continue
     no hang, device reboots.

Original-Change-Id: I9adbab3481dd77b315e1a987be24bf81d7c20d6e
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/304781
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Sean Paul <seanpaul@google.com>
Change-Id: I14d62ce7051eac6186cd70edb5a354ebfeb7a1bf
Reviewed-on: https://chromium-review.googlesource.com/306850
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>