FROMLIST: usb: dwc2: Avoid more calls to dwc2_core_reset()
Calls to dwc2_core_reset() are currently very slow, taking at least
150ms (possibly more).  It behooves us to take as many of these calls
out as possible.

It turns out that the calls in dwc2_fs_phy_init() and dwc2_hs_phy_init()
should (as documented in the code) only be needed if we need to do a PHY
SELECT.  That means that if we see that we can avoid the PHY SELECT then
we can avoid the reset.

This patch appears to successfully bypass two resets (one per USB
device) on rk3288-based ARM Chromebooks.

BUG=chrome-os-partner:46295
TEST=Both ports work OK

Conflicts:
    drivers/usb/dwc2/core.c
...due to missing (1f91b4c usb: dwc2: rename all s3c_* to dwc2_*)

Change-Id: I5990eda4efa570e6407422e22bf531e79ff43b2e
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/7355241/)
Reviewed-on: https://chromium-review.googlesource.com/304822
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>