UPSTREAM: wireless42: import the v4.2 wireless stack
This change copies the whole wireless stack from linux v4.2 in this
kernel. The newly created directories are suffixed with -4.2 to follow
the existing convention.

The stack doesn't build yet, it is a clean copy of the v4.2 content.

Adding the following directories:
	drivers/net/wireless-4.2
	include/wireless-4.2
	net/mac80211-4.2
	net/wireless-4.2

Sequence to generate was:
	git clone git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
	cd linux
	git pull
	git checkout v4.2
	rsync -av drivers/net/wireless/ $CHROOT/.../kernel/v3.14/drivers/net/wireless-4.2
	rsync -av net/wireless/ $CHROOT/.../kernel/v3.14/net/wireless-4.2
	rsync -av net/mac80211/ $CHROOT/.../kernel/v3.14/net/mac80211-4.2

The include/wireless-4.2 files were copied individually from the same
linux-2.6 source tree. Initially populated with the same files that are
in include/wireless-3.18.

BUG=chromium:543285
TEST=none

Change-Id: Ic69688d9e022059339b73cb38c27e378edfee123
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/304864
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>