Revert "CHROMIUM: usb: dwc2: Make bus suspend/resume do something for dwc2"
This reverts commit c20c0df75d9f700b80cc3e0d72b0337c3c6887c4.

The idea here is to revert the commit so we can do a bunch of upstream
cherry-picks (because the cherry-picks touched this code a lot), then we
want to re-apply this CL.

See CL:306265 for the reapply.

Note that since this revert doesn't really seem to hurt anything,
not doing CQ-DEPEND on the reapply.  We'll get it eventually.

BUG=chrome-os-partner:46308
TEST=w/ series both USB ports are happy

Change-Id: I6d4977501799f2acfc7696ca3f7b8df2d18d9a8e
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/306223
Reviewed-by: liangfeng wu <wulf@rock-chips.com>
Reviewed-by: yunzhi li <lyz@rock-chips.com>